+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Helicopter RPM indicator with configurable limits
-- Target IAS indicator
+- Target IAS and VSI indicator
 - Traffic display on HSI
 
 ## [0.8.1] - 2023-04-17